Add prod and dev scripts for GTM
